Updated the aws test cases
