+++ internal/cloudproviders/aws/main_unit_test.go
@@ -1,0 +1,57 @@
+package aws
+
+import (
+	"context"
+	"fmt"
+	localstate "github.com/ksctl/ksctl/internal/storage/local"
+	"github.com/ksctl/ksctl/pkg/helpers/consts"
+	"github.com/ksctl/ksctl/pkg/logger"
+	"github.com/ksctl/ksctl/pkg/types"
+	storageTypes "github.com/ksctl/ksctl/pkg/types/storage"
+	"os"
+	"path/filepath"
+	"testing"
+)
+
+var (
+	fakeClientHA *AwsProvider
+	storeHA      types.StorageFactory
+
+	fakeClientManaged *AwsProvider
+	storeManaged      types.StorageFactory
+	parentCtx         context.Context
+	fakeClientVars    *AwsProvider
+	storeVars         types.StorageFactory
+
+	parentLogger types.LoggerFactory = logger.NewStructuredLogger(-1, os.Stdout)
+
+	dir = filepath.Join(os.TempDir(), "ksctl-aws-test")
+)
+
+func TestMain(m *testing.M) {
+
+	parentCtx = context.WithValue(
+		context.TODO(),
+		consts.KsctlCustomDirLoc,
+		dir)
+
+	fakeClientVars, _ = NewClient(parentCtx, types.Metadata{
+		ClusterName: "demo",
+		Region:      "fake-region",
+		Provider:    consts.CloudAws,
+		IsHA:        true,
+	}, parentLogger, &storageTypes.StorageDocument{}, ProvideClient)
+
+	storeVars = localstate.NewClient(parentCtx, parentLogger)
+	_ = storeVars.Setup(consts.CloudAws, "fake-region", "demo", consts.ClusterTypeHa)
+	_ = storeVars.Connect()
+
+	exitVal := m.Run()
+
+	fmt.Println("Cleanup..")
+	if err := os.RemoveAll(dir); err != nil {
+		panic(err)
+	}
+
+	os.Exit(exitVal)
+}
